refactor(api): split the chats grab-bag into feature-directory modules
src/chats/ had accreted five features; only some had modules. Now one
directory per feature, each a real NestJS module:

- src/runs/ — RunsModule (read surface), controller, repository, dto,
  delta-buffer
- src/compaction/ — CompactionModule (provides + exports
  CompactionService)
- src/titles/ — TitlesModule (provides + exports TitleService)
- src/chats/ keeps the loop, context builder, repositories, telemetry;
  ChatsModule imports CompactionModule/TitlesModule instead of
  re-providing their services

Deliberate boundary: compaction/titles/runs are all importable without
the chat HTTP surface — the exact shape the worker split (#50) needs.
Pure file moves + import/module rewiring; no behavior change.

Verified: tsgo + oxlint + prettier clean, unit 117, nest build, full
rls-test.sh green (RLS 11, queue 7, e2e 27 — the e2e boots the full
AppModule, proving the module graph live).
