+++ CHANGELOG.md
@@ -2,6 +2,7 @@ _Reverse-chronological record of shipped work — features, fixes, and chores. N
 
 # 2026-07-03
 
+- Reorganized `apps/api` into feature-directory modules: `runs/`, `compaction/`, and `titles/` move out of the `chats/` grab-bag into their own directories with real NestJS modules (`RunsModule` read surface, `CompactionModule`, `TitlesModule` — each importable without dragging the chat HTTP surface along, which is exactly what the worker split (#50) needs); `chats/` keeps the loop, context builder, repositories, and telemetry.
 - Documented horizontal scaling ([docs/scaling.md](docs/scaling.md)): the api×N/worker×M/single-Postgres topology, the invariants that keep replica scaling correct (stateless api, terminal-status-implies-terminal-event, RLS-in-DB), and the six design constraints the worker split (#48/#50) must respect — transactional enqueue via pg-boss's external-transaction support, per-chat FIFO via partial unique index + `key_strict_fifo`, worker concurrency for IO-bound runs, LISTEN/NOTIFY for live deltas (polling is the resume path only), `model.delta` retention, and the deadman sweep appending `run.expired`.
 - Landed the durable-run substrate (#48, first slice): `runs` and append-only `run_events` tables (SPEC §9.3–§9.4) with RLS `ENABLE`+`FORCE` and cross-tenant read/write denial proven live. Every user message now creates a run **in the same transaction** as the message, and the streaming loop dual-writes an ordered lifecycle log (`run.created` → `run.started` → `model.requested` → `model.completed` → `run.completed`/`run.failed`/`run.cancelled`) — the durable source of truth the SSE replay (#49) will read. Still to come in #48: the worker consuming from pg-boss, token-delta events, cancellation/heartbeat/timeout, and per-chat single-flight (deliberately deferred until heartbeat exists, since without it a crashed run would deadlock its chat).
 - Made durable runs observable and replayable (#48/#49 API side): the loop now persists coalesced `model.delta` events (size-buffered via a pure delta-buffer, ordered by a sequential write chain), and a new run read surface exposes `GET /api/v1/runs/:id` plus the SPEC §9.4 cursor SSE `GET /api/v1/runs/:id/events?after_sequence=N` — each frame's SSE `id:` is its event sequence, an in-flight run is polled until terminal, a finished run streams its tail and closes, and a reconnect resumes from the last id with nothing lost. Cross-tenant reads 404 on both endpoints (proven in e2e). The `apps/web` resume-on-refresh client remains open in #49.
