+++ apps/api/AGENTS.md
@@ -10,7 +10,7 @@ NestJS 11 backend: API + services, and owner of the database schema/migrations. 
 
 ## Structure
 
-- `src/` — feature modules, each as controller + service + module (`chats/`, `users/`)
+- `src/` — one directory per feature, each a NestJS module (`chats/`, `runs/`, `compaction/`, `titles/`, `queue/`, `models/`, `auth/`, `users/`); a feature that another feature consumes exports its service from its own module (e.g. `ChatsModule` imports `CompactionModule`/`TitlesModule`), it is not re-provided elsewhere
 - `src/db/` — `schema/` (`auth.ts`, `chats.ts`), `migrations/` (+ `meta/` journal), `migrate.ts`
 - `src/main.ts`, `src/app.module.ts`
